part4 exercises 4.18-4.21

+++ part4/bloglist/controllers/blogs.js
@@ -1,3 +1,4 @@
+const jwt = require('jsonwebtoken');
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
@@ -18,8 +19,11 @@ blogsRouter.get('/:id', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body;
-
-    const user = await User.findById(body.userId);
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken.id) {
+        return response.status(401).json({error: 'token missing or invalid'});
+    }
+    const user = await User.findById(decodedToken.id);
 
     const blog = new Blog({
         title: body.title,
@@ -37,8 +41,20 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+    if (!decodedToken.id) {
+        return response.status(401).json({error: 'token missing or invalid'});
+    }
+
+    const user = await User.findById(decodedToken.id);
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id);
+        response.status(204).end();
+    } else {
+        response.status(401).json({error: 'You can not delete other users\' blogs'});
+    }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -51,7 +67,7 @@ blogsRouter.put('/:id', async (request, response) => {
         likes: body.likes
     };
 
-    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true });
+    const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true});
     response.status(201).json(updatedBlog);
 });
